Deterministic wiggle and new shuffle property for TweenSpace.sequential() method.
The wiggle effect is now deterministic thanks to the linear congruential
generator algorithm. In order to get different random results just
change the seed property value at instantiation time.
TweenSpace.sequential() as well as TweenSpace.sequentialTo() has two new
properties, shuffle and seed. If shuffle is set to true, the tweens are
going to be played in a deterministic random fashion. Change the seed
value in order to get different random behaviors.