TweenSpace evolves!
TweenSpace started as an in-house solution for tweening basic html
properties and managing conflicts on multiple asynchronous calls. Now,
this library has grown significantly and therefore some changes have
occurred in this build in order to keep the project organized and
maintainable. This build is fully backwards compatible, so no changes to
your existing code are needed. The modules have been separated into
different files and folders for better readability and extensibility to
future modules.